Very simple start to yet another Laravel CRUD/Admin generator. I hope to make this one suck less than the others.
Start with just a command to generate empty controllers and append a route.
Decided to use Blade as the template instead of stub files to have the full power of a template engine.
Only problem is that I need to remove the <?php tag for it to not be parsed as a normal php file.